+++ packages/qwik-city/middleware/node/http.ts
@@ -4,6 +4,23 @@ import type {
   ServerRequestEvent,
 } from '@builder.io/qwik-city/middleware/request-handler';
 
+const { ORIGIN, PROTOCOL_HEADER, HOST_HEADER = 'host' } = process.env;
+
+function getOrigin(req: IncomingMessage) {
+  const headers = req.headers;
+  const protocol =
+    (PROTOCOL_HEADER && headers[PROTOCOL_HEADER]) ||
+    ((req.socket as any).encrypted || (req.connection as any).encrypted ? 'https' : 'http');
+  const host = headers[HOST_HEADER];
+
+  return `${protocol}://${host}`;
+}
+
+export function getUrl(req: IncomingMessage) {
+  const origin = ORIGIN ?? getOrigin(req);
+  return new URL(req.url || '/', origin);
+}
+
 export async function fromNodeHttp(
   url: URL,
   req: IncomingMessage,
